requirements.txt
alembic==1.17.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.12.0
backoff==2.2.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
clickhouse-connect==0.10.0
clickhouse-driver==0.2.10
clickhouse-migrate==0.0.4
coloredlogs==14.0
cryptography==46.0.3
dagster==1.12.6
dagster-cloud==1.12.6
dagster-cloud-cli==1.12.6
dagster-graphql==1.12.6
dagster-msteams==0.28.6
dagster-pipes==1.12.6
dagster-polars==0.27.8
dagster-webserver==1.12.6
-e git+https://github.com/dagster-io/dagster-quickstart@089e7091a08ec1be97c9f43aad629c68f7d71270#egg=dagster_quickstart
dagster_shared==1.12.6
docstring_parser==0.17.0
envparse==0.2.0
exchange_calendars==4.11.3
filelock==3.20.1
fsspec==2025.12.0
github3.py==4.0.1
googleapis-common-protos==1.72.0
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-health-checking==1.76.0
h11==0.16.0
httptools==0.7.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
Jinja2==3.1.6
korean-lunar-calendar==0.3.1
lz4==4.4.5
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.0
pandas_market_calendars==5.2.2
pathlib_abc==0.5.2
pex==2.59.5
platformdirs==4.5.1
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyluach==2.3.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-stdnum==2.1
pytz==2025.2
PyYAML==6.0.3
questionary==2.1.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
structlog==25.5.0
tabulate==0.9.0
tomli==2.3.0
tomlkit==0.13.3
toolz==1.1.0
toposort==1.10
tqdm==4.67.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
universal_pathlib==0.3.7
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
validators==0.35.0
watchdog==5.0.3
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
xbbg==0.9.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
